Bug: Fixed text input auto selection on iOS Safari
A major annoyance is that when trying to enter the score on Mobile Safari, iOS, it was hard to get rid of the default value "-" as text is not selected on focus (as it is on other browsers/devices).

This is kind of a hack, but it solves the issue. t's a known bug in mobile safari:
http://stackoverflow.com/questions/3272089/programmatically-selecting-text-in-an-input-field-on-ios-devices-mobile-safari

On a side note: This file is full of mix between tabs and spaces, and I stuck with tabs because there seems to be no default convention. Should probably be fixed in separate PR/commit