Teach pdf-extract.py to update time and court
When reading from the .ods files, the time and court are unfortunately
mixed up. The error is actually in the .ods files provided by the
club. They adjust the time and court in the Access database when
connecting to the booking system. What's provided in the .pdf files is
correct, so teach pdf-extracy.py to update the database after-the-fact
to be able to use that data.

Importing from .ods files has other advantages, like names not being
trunctated so now it's recommended to import from .ods first and
update from the .pdf second.